requirements.txt
Flask
flask-cors
requests
beautifulsoup4
python-dotenv
